+++ R/run_cardoon.R
@@ -10,6 +10,8 @@
 #' @param sleep_time integer with the number of seconds the background process
 #'   sleeps.
 #' @param docs a boolean indicating if the docs should be started.
+#' @param log_path a path to a file where the logs of the backend
+#'  process are stored
 #'
 #' @import plumber
 #' @import logger
@@ -18,14 +20,22 @@
 #' @export
 #'
 run_cardoon <- function(
-    port = 9662,
-    num_worker = 1,
-    check_seconds = 60,
-    sleep_time = 10,
-    docs = FALSE) {
+  port = 9662,
+  num_worker = 1,
+  check_seconds = 10,
+  sleep_time = 5,
+  docs = FALSE,
+  log_path = "logs/"
+  ) {
 
-  # TODO add logging wihtin API on different levels (info, debug, ...)
+  logger::log_info("caRdoon version ", as.character(packageVersion("caRdoon")), "\n")
+
+  # TODO add logging within API on different levels (info, debug, ...)
+  # TODO add timestamp to logfile per default
+  # TODO use file.path()
   # TODO check put this blog for plumber package structure
+  # TODO check what happens to the background process if API process is killed
+
   # https://community.rstudio.com/t/plumber-api-and-package-structure/18099/11
 
   logger::log_info("set env vars for caRdoon API")
@@ -35,6 +45,8 @@ run_cardoon <- function(
   Sys.setenv(CARDOON_NUM_WORKER = num_worker)
   Sys.setenv(CARDOON_CHECK_SECONDS = check_seconds)
   Sys.setenv(CARDOON_SLEEP_TIME = sleep_time)
+  Sys.setenv(CARDOON_LOG_PATH = log_path)
+
 
   logger::log_info("start caRdoon API")
   plumber::plumb_api(package = "caRdoon", name = "cardoon") %>%
@@ -43,9 +55,10 @@ run_cardoon <- function(
       # naming is done later in plumber.R
       handler =  function() {
         logger::log_info("closing DB ...")
-        #DBI::dbDisconnect(cardoon_db)
+        # DBI::dbDisconnect(cardoon_db)
         logger::log_info("closing DB done")
-        }) %>%
+      }
+    ) %>%
     plumber::pr_run(
       # manually set port
       port = port,
@@ -57,4 +70,3 @@ run_cardoon <- function(
   logger::log_info("end caRdoon API")
   return("API closed")
 }
-